Refactor systemstatsmonitor/metric_helper.go into a metrics package
