+++ pkg/systemstatsmonitor/system_stats_monitor.go
@@ -22,6 +22,7 @@ import (
 	"time"
 
 	"github.com/golang/glog"
+
 	"k8s.io/node-problem-detector/pkg/problemdaemon"
 	ssmtypes "k8s.io/node-problem-detector/pkg/systemstatsmonitor/types"
 	"k8s.io/node-problem-detector/pkg/types"
